requirements.txt
PyQt5
tensorflow
keras
numpy
matplotlib
pillow
scikit-learn
imutils
opencv-python